+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add form helpers for security key registration and 2FA authentication.
 - Fix incorrect call to `resource_name` instead of using passed `resource` param in `login_with_security_key_button` helper.
 - Fix `NoMethodError` when calling `second_factor_enabled?` on resources without 2FA.
+- Avoid assuming `email` as the authentication key of the resource in form helpers.
 
 ## [v0.2.0](https://github.com/cedarcode/devise-webauthn/compare/v0.1.2...v0.2.0/) - 2025-12-03